Refactor Uber Command
