+++ app/controllers/concerns/uber_command.rb
@@ -1,64 +1,11 @@
 require 'addressable/uri'
-
-BASE_URL = ENV["uber_base_url"]
-
-# Leave out 'products' until user can pick.
-VALID_COMMANDS = [
-  'ride',
-  'estimate',
-  'help',
-  'accept',
-  'share',
-  'status',
-  'cancel',
-  'trigger_error',
-  'test_resque'
-]
-
-# returned when ride isn't requested in the format '{origin} to {destination}'
-RIDE_REQUEST_FORMAT_ERROR = <<-STRING
-  To request a ride please use the format */uber ride [origin] to [destination]*.
-  For best results, specify a city or zip code.
-  Ex: */uber ride 1061 Market Street San Francisco to 55 Music Concourse Dr San Francisco*
-STRING
-
-ESTIMATES_FORMAT_ERROR = <<-STRING
-  To request estimates for a trip, please use the format */uber [origin] to [destination]*.
-  For best results, specify a city or zip code.
-  Ex: */uber estimate 1061 Market Street San Francisco to 55 Music Concourse Dr San Francisco*
-STRING
-
-UNKNOWN_COMMAND_ERROR = <<-STRING
-  Sorry, we didn't quite catch that command.  Try */uber help* for a list.
-STRING
-
-# Products is left out
-HELP_TEXT = <<-STRING
-  Try these commands:
-  - ride [origin address] to [destination address]
-  - estimate [origin address] to [destination address]
-  - share
-  - status
-  - cancel
-  - help
-
-  For best results when requesting a ride or estimate, specify a city or zip code.
-  Ex: */uber ride 1061 Market Street, San Francisco to 55 Music Concourse Dr, San Francisco*
-STRING
-
-RIDE_STATUSES = {
-  "processing" => "Looking for a driver.",
-  "no_drivers_available" => "No drivers were available to pick you up. Try again.",
-  "accepted" => "A driver accepted your request and is on their way.",
-  "arriving" => "Your driver is arriving now.",
-  "in_progress" => "Your ride is in progress.",
-  "driver_canceled" => "Your driver canceled.",
-  "rider_canceled" => "You canceled the last ride you requested through Slack.",
-  "completed" => "You completed the last ride you requested through Slack."
-}
+require 'uber_command_error_strings'
+class FormatError < StandardError; end
 
 class UberCommand
+  include UberCommandFormatters
 
+  attr_reader :bearer_token
   def initialize(bearer_token, user_id, response_url)
     @bearer_token = bearer_token
     @user_id = user_id
@@ -74,29 +21,22 @@ class UberCommand
 
     command_argument = input.second.nil? ? "" : input.second.downcase
 
-    return UNKNOWN_COMMAND_ERROR if invalid_command?(command_name) || command_name.nil?
+    if invalid_command?(command_name) || command_name.nil?
+      return UNKNOWN_COMMAND_ERROR
+    end
 
-    self.send(command_name, command_argument)
+    begin
+      self.send(command_name, command_argument)
+    rescue FormatError => e
+      return e.message
+    end
   end
 
   private
 
-  attr_reader :bearer_token
-
   def estimate(user_input_string)
-    return ESTIMATES_FORMAT_ERROR unless user_input_string.include?(" to ")
-
-    start_addr, end_addr = parse_start_and_end_address(user_input_string)
-    start_lat, start_lng = resolve_address(start_addr)
-    end_lat, end_lng = resolve_address(end_addr)
-
-    return ["Sorry, we don't know where #{start_addr} is.",
-            "Can you try again with a more precise origin address?"
-    ].join(" ") if start_lat.nil?
-
-    return ["Sorry, we don't know where #{end_addr} is.",
-            "Can you try again with a more precise destination address?"
-    ].join(" ") if end_lat.nil?
+    origin_lat, origin_lng, destination_lat, destination_lng =
+      parse_start_and_end_coords(user_input_string, ESTIMATES_FORMAT_ERROR)
 
     product_id = get_default_product_id_for_lat_lng(start_lat, start_lng)
       return [
@@ -105,13 +45,15 @@ class UberCommand
       ].join(" ") if product_id.nil?
 
     begin
-      ride_estimate_hash = get_ride_estimate(
-        start_lat,
-        start_lng,
-        end_lat,
-        end_lng,
-        product_id
-    )
+      body = {
+        "start_latitude" => origin_lat,
+        "start_longitude" => origin_lng,
+        "end_latitude" => destination_lat,
+        "end_longitude" => destination_lng,
+        "product_id" => product_id
+      }
+
+      ride_estimate_hash = UberAPI.get_ride_estimate(body, bearer_header)
     rescue => e
       Rollbar.error(e, "UberCommand#estimate")
       return [
@@ -129,10 +71,7 @@ class UberCommand
 
   def share(_) # No command argument.
     ride = Ride.where(user_id: @user_id).order(:updated_at).last
-
-    if ride.nil?
-      return "Sorry, we couldn't find a ride for you to share."
-    end
+    return "Sorry, we couldn't find a ride for you to share." if ride.nil?
 
     request_id = ride.request_id
 
@@ -152,12 +91,9 @@ class UberCommand
     "Use this link to share your ride's progress: #{link}."
   end
 
-  def status(_)# No command argument.
+  def status(_) # No command argument.
     ride = Ride.where(user_id: @user_id).order(:updated_at).last
-
-    if ride.nil?
-      return "Sorry, we couldn't find any rides that you requested."
-    end
+    return "Sorry, we couldn't find any rides that you requested." if ride.nil?
 
     begin
       status_hash = get_ride_status(ride.request_id)
@@ -169,19 +105,15 @@ class UberCommand
     ride_status = status_hash["status"]
 
     eta = status_hash["eta"]
+    eta_msg = eta ? "ETA: #{eta} minutes" : nil
+    eta_msg = "ETA: one minute" if eta == 1
 
-    case eta
-    when nil then eta_msg = nil
-    when 1 then eta_msg = "ETA: one minute"
-    else eta_msg = "ETA: #{eta} minutes"
-    end
-
-    if [
-      "processing",
-      "accepted",
-      "arriving",
-      "in_progress"
-    ].include?(ride_status)
+    if %w(
+      processing
+      accepted
+      arriving
+      in_progress
+    ).include?(ride_status)
       return [
         "STATUS:",
         RIDE_STATUSES[ride_status],
@@ -189,20 +121,19 @@ class UberCommand
       ].compact.join(" ")
     end
 
-    return "STATUS: #{RIDE_STATUSES[ride_status]}"
+    "STATUS: #{RIDE_STATUSES[ride_status]}"
   end
 
   def cancel(_) # No command argument.
     ride = Ride.where(user_id: @user_id).order(:updated_at).last
-    if ride.nil?
-      return "Sorry, we couldn't find a ride for you to cancel."
-    end
+    return "Sorry, we couldn't find a ride for you to cancel." if ride.nil?
 
     request_id = ride.request_id
 
     fail_msg = "Sorry, we were unable to cancel your ride."
 
     begin
+      UberAPI.cancel_ride(request_id, bearer_header)
       resp = RestClient.delete(
         "#{BASE_URL}/v1/requests/#{request_id}",
         authorization: bearer_header,
@@ -215,7 +146,7 @@ class UberCommand
     end
 
     return "Successfully canceled your ride." if resp.code == 204
-    return fail_msg
+    fail_msg
   end
 
   def get_ride_status(request_id)
@@ -304,21 +235,9 @@ class UberCommand
     end
   end
 
-  def ride(input_str)
-    return RIDE_REQUEST_FORMAT_ERROR unless input_str.include?(" to ")
-
-    origin_name, destination_name = parse_start_and_end_address(input_str)
-    origin_lat, origin_lng = resolve_address origin_name
-    destination_lat, destination_lng = resolve_address destination_name
-
-    return ["Sorry, we don't know where #{origin_name} is.",
-            "Can you try again with a more precise origin address?"
-    ].join(" ") if origin_lat.nil?
-
-    return ["Sorry, we don't know where #{destination_name} is.",
-            "Can you try again with a more precise destination address?"
-    ].join(" ") if destination_lat.nil?
-
+  def ride(user_input_string)
+    origin_lat, origin_lng, destination_lat, destination_lng =
+      parse_start_and_end_coords(user_input_string, RIDE_REQUEST_FORMAT_ERROR)
 
     product_id = get_default_product_id_for_lat_lng(origin_lat, origin_lng)
     return [
@@ -326,14 +245,16 @@ class UberCommand
       "Can you try again with a more precise address?"
     ].join(" ") if product_id.nil?
 
+    body = {
+      "start_latitude" => origin_lat,
+      "start_longitude" => origin_lng,
+      "end_latitude" => destination_lat,
+      "end_longitude" => destination_lng,
+      "product_id" => product_id
+    }
+
     begin
-      ride_estimate_hash = get_ride_estimate(
-        origin_lat,
-        origin_lng,
-        destination_lat,
-        destination_lng,
-        product_id
-      )
+      ride_estimate_hash = UberAPI.get_ride_estimate(body, bearer_header)
     rescue => e
       Rollbar.error(e, "UberCommand#ride")
       return [
@@ -383,145 +304,10 @@ class UberCommand
     "#{ack} Pinging Uber to drive you from #{origin_name} to #{destination_name}..."
   end
 
-  def ask_for_surge_confirmation(multiplier)
-    base = "#{multiplier}x surge is in effect."
-
-    if multiplier >= 2
-      [base, "Reply */uber accept #{multiplier}'* to confirm the ride."].join(" ")
-    else
-      [base, "Reply */uber accept* to confirm the ride"].join(" ")
-    end
-  end
-
-  def parse_start_and_end_address(input_str)
-    origin_name, destination_name = input_str.split(" to ")
-
-    if origin_name.start_with? "from "
-      origin_name = origin_name["from".length..-1]
-    end
-
-    origin_name = origin_name.strip
-    destination_name = destination_name.strip
-
-    [origin_name, destination_name]
-  end
-
-  def get_ride_estimate(start_lat, start_lng, end_lat, end_lng, product_id)
-    body = {
-      "start_latitude" => start_lat,
-      "start_longitude" => start_lng,
-      "end_latitude" => end_lat,
-      "end_longitude" => end_lng,
-      "product_id" => product_id
-    }
-
-    response = RestClient.post(
-      "#{BASE_URL}/v1/requests/estimate",
-      body.to_json,
-      authorization: bearer_header,
-      "Content-Type" => :json,
-      accept: :json
-    )
-
-    JSON.parse(response.body)
-  end
-
   def reply_to_slack(response)
-      payload = { text: response }
+    payload = { text: response }
 
-      RestClient.post(@response_url, payload.to_json)
-  end
-
-  def get_default_product_id_for_lat_lng(lat, lng)
-    product_id = Rails.cache.fetch("location: #{lat}/#{lng}", expires_in: 15.minutes) do
-      available_products = get_products_for_lat_lng(lat, lng)["products"]
-      available_products.empty? ? nil : available_products.first["product_id"]
-    end
-
-    product_id
-  end
-
-  def get_products_for_lat_lng(lat, lng)
-    uri = Addressable::URI.parse("#{BASE_URL}/v1/products")
-    uri.query_values = { 'latitude' => lat, 'longitude' => lng }
-    resource = uri.to_s
-
-    result = RestClient.get(
-      resource,
-      authorization: bearer_header,
-      "Content-Type" => :json,
-      accept: 'json'
-    )
-
-    JSON.parse(result.body)
-  end
-
-  def format_200_ride_request_response(origin, destination, response)
-    ack = ["Got it!", "Roger that.", "OK.", "10-4."].sample
-
-    ["#{ack} We are looking for a driver",
-     "to take you from #{origin} to #{destination}.",
-    ].join(" ")
-  end
-
-  def format_response_errors(response_errors)
-    response = "The following errors occurred: \n"
-    response_errors.each do |error|
-      response += "- *#{error['title']}* \n"
-    end
-  end
-
-  def format_products_response(products_response)
-    unless products_response['products'] && !products_response['products'].empty?
-      return "No Uber products available for that location."
-    end
-    response = "The following products are available: \n"
-    products_response['products'].each do |product|
-      response += "- #{product['display_name']}: #{product['description']} (Capacity: #{product['capacity']})\n"
-    end
-    response
-  end
-
-  def format_ride_estimate_response(start_addr, end_addr, ride_estimate_hash)
-    duration_secs = ride_estimate_hash["trip"]["duration_estimate"]
-    duration_mins = duration_secs / 60
-
-    duration_msg = duration_mins == 1 ? "one minute" : "#{duration_mins} minutes"
-
-    cost = ride_estimate_hash["price"]["display"]
-    surge = ride_estimate_hash["price"]["surge_multiplier"]
-    surge_msg = surge == 1 ? "No surge is currently in effect." : "Includes current surge at #{surge}x."
-
-    [
-      "Driving from #{start_addr} to #{end_addr} would take",
-      "about #{duration_msg} and cost #{cost}.",
-      surge_msg
-    ].join(" ")
-  end
-
-  def bearer_header
-    "Bearer #{bearer_token}"
-  end
-
-  def invalid_command?(name)
-    !VALID_COMMANDS.include? name
-  end
-
-  def resolve_address(address)
-    location = Rails.cache.fetch("address: #{address}", expires_in: 1.day) do
-      Geocoder.search(address).first
-    end
-
-    if location.blank?
-      return [nil, nil]
-    else
-      location = location.data["geometry"]["location"]
-      [location['lat'], location['lng']]
-    end
-  end
-
-  def trigger_error(_) # No command argument.
-    fail
+    RestClient.post(@response_url, payload.to_json)
   end
 
   def test_resque(input)
